requirements.txt
flask
psycopg2-binary
pandas
pyserial
sqlalchemy
fastapi
uvicorn
sqlalchemy
pydantic
databases
dotenv